requirements.txt
fastapi
uvicorn[standard]
boto3
python-multipart

opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-urllib3
